cloud_run/paper_discovery/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-bigquery>=3.12.0
google-cloud-storage>=2.10.0
arxiv>=1.4.8
pydantic>=2.5.0
asyncio-throttle>=1.0.2
tenacity>=8.2.3
structlog>=23.2.0
python-multipart>=0.0.6
